llm/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-docx==1.2.0
python-pptx==1.0.2
PyPDF2==3.0.1
openai==1.90.0
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-community==0.3.26
langchain-text-splitters==0.3.8
sentence-transformers==4.1.0
pydantic==2.11.7
pydantic-settings==2.10.0
typing_extensions==4.14.0
requests==2.32.4
tiktoken==0.9.0
numpy==2.3.1
aiohttp==3.12.13
httpx==0.28.1
python-dotenv==1.1.0
pinecone-client==3.0.2
